refactor(auth): move cookie pair loading to leaf
