+++ src/notebooklm/_auth/cookie_contract.py
@@ -1,0 +1,5 @@
+"""Dependency-bottom wire contracts shared by cookie routing and recovery."""
+
+from __future__ import annotations
+
+KEEPALIVE_ROTATE_URL = "https://accounts.google.com/RotateCookies"
